Removed old resolver code.
The new and old code was tangled up a little, I wanted to keep it for historical purposes, however it's just too messy and we should always be aiming to delete as much code as we can. No code is the best code.
